gitignore

+++ FCFS.java
@@ -19,17 +19,22 @@ public class FCFS {
             p.tempoDeEspera = tempoAtual - p.tempoDeChegada;
             p.tempoDeRetorno = p.tempoDeEspera + p.tempoDePico;
             tempoAtual += p.tempoDePico;
+            //System.out.println("tempoAtual = " + tempoAtual);
         }
     }
 
     public void printarMedias() {
         int totalRetorno = 0, totalResposta = 0, totalEspera = 0;
         for (Processo p : fila) {
+            //System.out.println("ret = " + p.tempoDeRetorno);
+            //System.out.println("res = " + p.tempoDeResposta);
+            //System.out.println("esp = " + p.tempoDeEspera);
+
             totalRetorno += p.tempoDeRetorno;
             totalResposta += p.tempoDeResposta;
             totalEspera += p.tempoDeEspera;
         }
-        System.out.println("FCFS " + ((float) totalRetorno) / fila.size() + " " + totalResposta/fila.size() + " " + totalEspera/fila.size());
+        System.out.println("FCFS " + ((float) totalRetorno) / fila.size() + " " + ((float) totalResposta) / fila.size() + " " + ((float)totalEspera) / fila.size());
     }
     
 }
